Add reminder section to EditEvent.
Added reminder section to Web/Calendars/EditEvent.cs.
