+++ KinaUnaWeb/Scripts/calendar/calendar-index.ts
@@ -112,6 +112,29 @@ function addScheduleEventListeners(): void {
 }
 
 /**
+ * Shows the event details popup when the page is loaded, if the url query string contains an eventId.
+ */
+function showPopupAtLoad() {
+    const popupEventIdDiv = document.querySelector<HTMLDivElement>('#popup-event-id-div');
+    if (popupEventIdDiv !== null) {
+        if (popupEventIdDiv.dataset.popupEventId) {
+            let eventId = parseInt(popupEventIdDiv.dataset.popupEventId);
+            if (eventId > 0) {
+                if (popupEventIdDiv.dataset.popupEventDateYear && popupEventIdDiv.dataset.popupEventDateMonth && popupEventIdDiv.dataset.popupEventDateDay) {
+                    const popupEventYear = parseInt(popupEventIdDiv.dataset.popupEventDateYear);
+                    const popupEventMonth = parseInt(popupEventIdDiv.dataset.popupEventDateMonth) -1;
+                    const popupEventDay = parseInt(popupEventIdDiv.dataset.popupEventDateDay);
+                    let scheduleInstance = document.querySelector<any>('.e-schedule').ej2_instances[0];
+                    scheduleInstance.selectedDate = new Date(popupEventYear, popupEventMonth, popupEventDay);
+                }
+                
+                DisplayEventItem(eventId);
+
+            }
+        }
+    }
+}
+/**
  * Initializes page elements when it is loaded.
  */
 document.addEventListener('DOMContentLoaded', async function (): Promise<void> {
@@ -119,6 +142,8 @@ document.addEventListener('DOMContentLoaded', async function (): Promise<void> {
     addScheduleEventListeners();
     await loadLocale();
     setLocale();
+
+    showPopupAtLoad();
 
     return new Promise<void>(function (resolve, reject) {
         resolve();
