+++ KinaUnaWeb/Views/Calendar/AddEvent.cshtml
@@ -93,6 +93,7 @@
                     <select class="selectpicker form-control-noline" data-style="btn-kinaunaselect" data-size="@Model.AccessLevelListEn.Count" asp-for="CalendarItem.AccessLevel" asp-items="@Model.AccessLevelListEn"></select>
                 </div>
             </div>
+            
             <div>
                 <div class="form-group">
                     <div class="col-auto control-label"></div>
