+++ KinaUnaWeb/Views/Calendar/ViewEvent.cshtml
@@ -38,6 +38,10 @@
             <div class="space-20"></div>
             <div align="left"><strong>@await LocaleManager.GetTranslation("Notes:", PageNames.Calendar, Model.LanguageId)</strong> @Model.CalendarItem.Notes</div>
             <div class="space-20"></div>
+            @if (Model.CalendarItem.StartTime.HasValue && Model.CalendarItem.StartTime.Value > TimeZoneInfo.ConvertTimeFromUtc(DateTime.UtcNow, TimeZoneInfo.FindSystemTimeZoneById(Model.CurrentUser.Timezone)))
+            {
+                await Html.RenderPartialAsync("_CalendarRemindersPartial", Model);
+            }
             @if (Model.IsCurrentUserProgenyAdmin)
             {
                 <div>
